+++ frontend/pages/data/rates.py
@@ -1,4 +1,5 @@
 import requests
+import json
 from config import base_url
 from typing import List, Dict, TypedDict
 from datetime import datetime
@@ -49,4 +50,33 @@ def rates_by_user_client_date(user_id: int, client_id: int, date: str) -> List[D
 def rate_update(user_id: int, client_id: int, new_amount: float):
     api = f"{base_url}/api/rates/?user_id={user_id}&client_id={client_id}&new_amount={new_amount}"
     response = requests.put(api)
+    return True
+
+
+def to_rate(
+    user_id: int,
+    client_id: int,
+    valid_from: str,
+    valid_to: str,
+    amount: float,
+    created_at: str,
+    updated_at: str,
+    is_active: bool,
+) -> bool:
+    data = Rate(
+        user_id=user_id,
+        client_id=client_id,
+        valid_from=valid_from,
+        valid_to=valid_to,
+        amount=amount,
+        created_at=created_at,
+        updated_at=updated_at,
+        is_active=True,
+    )
+    print(data)
+    response = requests.post(
+        f"{base_url}/api/rates",
+        data=json.dumps(dict(data)),
+        headers={"accept": "application/json", "Content-Type": "application/json"},
+    )
     return True
